Implement fast direct creation
